requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
tensorflow>=2.16.0
mlflow==2.10.0
dvc==3.42.0
pytest==8.0.0
httpx==0.26.0
numpy
pillow==10.2.0
scikit-learn==1.4.0
protobuf==4.25.3
flake8
prometheus-fastapi-instrumentator
